pyproject.toml
[project]
name = "django-event"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework==3.15.2",
    "django>=5.1.3",
    "ruff>=0.7.3",
    "drf-yasg>=1.21.8",
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 80
